Move campaign form to a model
